requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
